#3304 - XHR done toggle and warn notice for post "done" hits

+++ classes/cubecart.class.php
@@ -2123,10 +2123,13 @@ class Cubecart
         foreach ($GLOBALS['hooks']->load('class.cubecart.404') as $hook) include $hook;
 
         if(!empty($_SERVER['REQUEST_URI'])) {
+            if(substr($_SERVER['REQUEST_URI'],-6)=='_a=404') return false;
+            
             $GLOBALS['db']->delete('CubeCart_404_log', '`created` < DATE_SUB(NOW(), INTERVAL 90 DAY)');
-            $uri = htmlentities($_SERVER['REQUEST_URI'],ENT_QUOTES);
-            if($GLOBALS['db']->select('CubeCart_404_log', false, array('uri' => $uri))) {
-                $GLOBALS['db']->update('CubeCart_404_log', array('hits' => '+1'), array('uri' => $uri));
+            $uri = htmlentities(ltrim($_SERVER['REQUEST_URI'], CC_ROOT_REL),ENT_QUOTES);
+            if($existing = $GLOBALS['db']->select('CubeCart_404_log', false, array('uri' => $uri))) {
+                $warn = ($existing[0]['done'] == 1) ? 1 : 0;
+                $GLOBALS['db']->update('CubeCart_404_log', array('hits' => '+1', 'warn' => $warn), array('uri' => $uri));
             } else {
                 $GLOBALS['db']->insert('CubeCart_404_log', array('uri' => $uri));
             }
